+++ react-app/src/store/auth.js
@@ -1,44 +1,59 @@
-// Basic Skeleton auth.js im 99.99% sure we dont have to change anything in this file, including sign up.
+const SET_USER = "setUser";
+const REMOVE_USER = "removeUser";
 
-export const authenticate = async() => {
-  const response = await fetch('/api/auth/',{
-    headers: {
-      'Content-Type': 'application/json'
-    }
-  });
-  return await response.json();
-}
+const setUser = (user) => ({
+  type: SET_USER,
+  user,
+});
 
-export const login = async (email, password) => {
-  const response = await fetch('/api/auth/login', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json'
-    },
-    body: JSON.stringify({
-      email,
-      password
-    })
-  });
-  return await response.json();
-}
+const removeUser = () => ({
+  type: REMOVE_USER,
+});
 
-export const logout = async () => {
-  const response = await fetch("/api/auth/logout", {
+export const authenticate = async () => {
+  const response = await fetch("/api/auth/", {
     headers: {
       "Content-Type": "application/json",
-    }
+    },
   });
   return await response.json();
 };
 
-
-export const signUp = async (userName, firstName, lastName, email, password) => {
-  const response = await fetch("/api/auth/signup", {
+export const login = (email, password) => async (dispatch) => {
+  const build = {
     method: "POST",
-    headers: {
-      "Content-Type": "application/json",
-    },
+    headers: { "Content-Type": "application/json" },
+    body: JSON.stringify({
+      email,
+      password,
+    }),
+  };
+  const response = await fetch("/api/auth/login", build);
+  const user = await response.json();
+  dispatch(setUser(user));
+  return user;
+};
+
+export const logout = () => async (dispatch) => {
+  const build = {
+    headers: { "Content-Type": "application/json" },
+  };
+  const response = await fetch("/api/auth/logout", build);
+  dispatch(removeUser());
+  const res = await response.json();
+  return res;
+};
+
+export const signUp = (
+  userName,
+  firstName,
+  lastName,
+  email,
+  password
+) => async (dispatch) => {
+  const build = {
+    method: "POST",
+    headers: { "Content-Type": "application/json" },
     body: JSON.stringify({
       userName,
       firstName,
@@ -46,6 +61,29 @@ export const signUp = async (userName, firstName, lastName, email, password) => 
       email,
       password,
     }),
-  });
-  return await response.json();
-}
+  };
+  const response = await fetch("/api/auth/signup", build);
+  const user = await response.json();
+  dispatch(setUser(user));
+  return user;
+};
+
+const initialState = { user: {} };
+
+const authReducer = (state = initialState, action) => {
+  let newState;
+  switch (action.type) {
+    case SET_USER:
+      newState = Object.assign({}, state);
+      newState.user = action.payload;
+      return newState;
+    case REMOVE_USER:
+      newState = Object.assign({}, state);
+      newState.user = null;
+      return newState;
+    default:
+      return state;
+  }
+};
+
+export default authReducer;
